Implement Hand.strain, to rate difficulty of holding a given shape.
Worth noting that the open D should be strain 2 with my preferred
three-finger shape, but doing a half-barre with only two is totally
fine and does hit 0 strain, so I passed it. May be an issue later,
may just be a bad habit on my part. Either way, good job Mr. Hand()