Implement Guitarist.play, to navigate Song structures
For a basic scale, it prefers to slide straight up one string with the index
finger alone, rather than walking up a more typical box pattern with all
the fingers. This is expected. I will need to implement a means of
picking fingers to use preferentially to minimize strain within individual
Hand positions and movements.