Implement Guitarist.read, to reduce complexity of play
Since each chord's ideal shape is only directly impacted by its immediate
neighbors, there is no need to revisit early notes after moving on to
later parts of a song. Therefore, a song can be broken down into smaller
sections to be played separately, and an ideal arrangement can be
constructed from the results of these otherwise independent passages.

This reduces complexity of the 'Blackbird' test to a mere 812 variants,
which is completed on the order of milliseconds. Unfortunately the result
still differs from expected. At least it does so in an entirely sensible way.